处理一些问题。 端口配置

+++ model/model.go
@@ -15,7 +15,7 @@ type Peer struct {
 type Relay struct {
 	Uid         string     `json:"_id"`
 	Name        string     `json:"name"`
-	Port        uint32     `json:"port"`
+	Port        uint16     `json:"port"`
 	IP          string     `json:"ip"`
 	Online      bool       `json:"online"`
 	LastRegTime *time.Time `json:"last_reg_time"`
